Update backup utilities to log info messages only when command trace setting is on.

+++ skyve-ext/src/main/java/org/skyve/impl/backup/Reindex.java
@@ -27,7 +27,7 @@ public class Reindex {
 
 		// truncate the bean content ready to reindex
 		try (ContentManager cm = EXT.newContentManager()) {
-			UtilImpl.LOGGER.info("Truncate Beans");
+			if (UtilImpl.COMMAND_TRACE) UtilImpl.LOGGER.info("Truncate Beans");
 			cm.truncateBeans(customer.getName());
 		}
 		
@@ -49,7 +49,7 @@ public class Reindex {
 								// Don't check if a document has indexable fields as we
 								// may need to have nodes deleted
 								// (i.e. a document field used to be indexed but now is not)
-								UtilImpl.LOGGER.info(String.format("Reindex document %s.%s", module.getName(), documentName));
+								if (UtilImpl.COMMAND_TRACE) UtilImpl.LOGGER.info(String.format("Reindex document %s.%s", module.getName(), documentName));
 								DocumentQuery query = persistence.newDocumentQuery(document);
 								for (Bean bean : query.beanIterable()) {
 									persistence.reindex((PersistentBean) bean);
